Resolving comments and upgrading code
