+++ public/components/overview/mitre_attack_intelligence/module_mitre_attack_intelligence_resource.tsx
@@ -13,8 +13,9 @@
 
 import React, { useCallback, useState, useEffect } from 'react';
 import { TableWzAPI } from '../../../components/common/tables';
-import { ModuleMitreAttackIntelligenceFlyout } from './module_mitre_attack_intelligence_resource_flyout';
 import { WzRequest } from '../../../react-services';
+import { ModuleMitreAttackIntelligenceFlyout } from './resource_detail_flyout';
+
 
 export const ModuleMitreAttackIntelligenceResource = ({ label, searchBarSuggestions, apiEndpoint, tableColumns, initialSortingField, resourceFilters }) => {
